joget_validator/requirements.txt
mysql-connector-python==8.0.33
PyYAML==6.0